Fixed compiler warning.

+++ src/uiODMain/uiodviewer2dmgr.cc
@@ -363,7 +363,7 @@ void uiODViewer2DMgr::setVWR2DIntersectionPositions( uiODViewer2D* vwr2d )
 	    {
 		if ( idxvwrdir==TrcKeyZSampling::Crl )
 		{
-		    newpos.pos_ = idxvwrtkzs.hsamp_.crlRange().start;
+		    newpos.pos_ = (float) idxvwrtkzs.hsamp_.crlRange().start;
 		    newpos.name_ = tr( "CRL %1" ).arg( toString(newpos.pos_) );
 		    x1intposs += newpos;
 		}
